TgBot/requirements.txt
psycopg2-binary
pyTelegramBotAPI
pymorphy2
requests
